Move navigation to actions folder
