+++ examples/advanced/src/actions/navigation.ts
